backend/requirements.txt
# Krishna AI Backend Dependencies
# Multi-task Transformer + Whisper ASR

# Web framework
flask>=2.3.0
flask-cors>=4.0.0

# Deep learning
torch>=2.0.0
transformers>=4.30.0

# Speech-to-text
openai-whisper>=20231117

# Utilities
numpy>=1.24.0
